Used a common component for each tutorial/tour and refactor
